requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
APScheduler==3.10.4
python-dateutil==2.9.0.post0
matplotlib==3.9.0
reportlab==4.2.2
tzdata==2025.1
